fix: logger and update bt version
