+++ miner/pyproject.toml
@@ -7,7 +7,7 @@ readme = "../README.md"
 
 [tool.poetry.dependencies]
 python = ">=3.10,<3.13"
-bittensor = "9.0.0rc6"
+bittensor = "9.0.0"
 datatrove = {extras = ["all"], version = "^0.4.0"}
 nltk = "^3.9.1"
 tokenizers = "^0.21.0"
